Rework package tracking system (#104)
* Rework package tracking system

Numerous things have been changed/redone here, particularly the
way that packages are tracked as being installed.
Packages are now stored in a packages.list file located in
$VITASDK/etc/vdpm. This file keeps track of packages that are
installed and the files that each package has placed.
The /tmp-based package cache system has been removed.

Two new flags have been added; -f (force) and -u (uninstall).
-f will uninstall and reinstall a package, regardless of if
it's already installed. -u will simply uninstall a package.

The usage has also been modified to better reflect
available arguments.

* Safety fixes

Came up with a much better, actually safe implementation of the
package file deletion. Now, there is no filtration done on what
the archive placed in the packages.list file -- instead, checks
are run on each package entry that is scheduled for deletion.

In particular, if we do not have write access to the directory or
file, it is skipped. If a directory or file is wanted by another
package, then it will be skipped.

Also, if a directory is going to be deleted, it must be empty,
otherwise it will be skipped.

To prevent folders placed by the package being checked for removal
early, directories are now sorted & placed last in the
removal list.

No recursion of any kind happens here -- this should be completely
safe (without outside, intentional local tampering.)

* Correct comment inaccuracy